Function rust_implementation::lib::planet::visual_aspects_of_a_planet

pub fn visual_aspects_of_a_planet(
    lct_hour: f64, 
    lct_min: f64, 
    lct_sec: f64, 
    is_daylight_saving: bool, 
    zone_correction_hours: i32, 
    local_date_day: f64, 
    local_date_month: u32, 
    local_date_year: u32, 
    planet_name: String
) -> (f64, f64, f64, f64, f64, f64, f64, f64)

Calculate several visual aspects of a planet.

Arguments

lct_hour -- Local civil time, hour part.

lct_min -- Local civil time, minutes part.

lct_sec -- Local civil time, seconds part.

is_daylight_saving -- Is daylight savings in effect?

zone_correction_hours -- Time zone correction, in hours.

local_date_day -- Local date, day part.

local_date_month -- Local date, month part.

local_date_year -- Local date, year part.

planet_name -- Name of planet, e.g., "Jupiter"

Returns

distance_au -- Planet's distance from Earth, in AU.

ang_dia_arcsec -- Angular diameter of the planet.

phase -- Illuminated fraction of the planet.

light_time_hour -- Light travel time from planet to Earth, hour part.

light_time_minutes -- Light travel time from planet to Earth, minutes part.

light_time_seconds -- Light travel time from planet to Earth, seconds part.

pos_angle_bright_limb_deg -- Position-angle of the bright limb.

approximate_magnitude -- Apparent brightness of the planet.